+++ bssopenapi-20171214/src/Models/UnsubscribeBillToOSSResponse.php
@@ -30,9 +30,6 @@ class UnsubscribeBillToOSSResponse extends Model
 
     public function validate()
     {
-        Model::validateRequired('headers', $this->headers, true);
-        Model::validateRequired('statusCode', $this->statusCode, true);
-        Model::validateRequired('body', $this->body, true);
     }
 
     public function toMap()