+++ bssopenapi-20171214/ChangeLog.md
@@ -1,3 +1,9 @@
+2024-03-13 Version: 2.0.0
+- Delete API SaveUserCredit.
+- Delete API SetCreditLabelAction.
+- Update API SetRenewal: update param ProductCode.
+
+
 2023-12-13 Version: 1.2.1
 - Generated php 2017-12-14 for BssOpenApi.